requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
WTForms>=3.1.1
Werkzeug>=3.0.1
email-validator>=2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# Production server
gunicorn>=21.2.0
waitress>=3.0.0
requests>=2.31.0
redis>=5.0.0

# Optional: PostgreSQL
psycopg2-binary>=2.9.9
Flask-Limiter>=3.5.0
